Adding a New Layout and Styles to the Page
A `Layout.cshtml` file has been created that defines the basic structure of the page, including the header, body, and footer, along with a navigation menu.

_ViewStart.cshtml` has been updated to use the new layout.

Styles have been added to `layout.css` for the menu and footer layout, as well as for the main container.

In addition, CSS variables have been defined in `properties.css` for more consistent style management.